+++ packages/vm/src/testing/mock.rs
@@ -33,12 +33,14 @@ pub fn mock_backend_with_balances(
     }
 }
 
-/// Length of canonical addresses created with this API. Contracts should not make any assumtions
+/// Length of canonical addresses created with this API. Contracts should not make any assumptions
 /// what this value is.
+///
 /// The value here must be restorable with `SHUFFLES_ENCODE` + `SHUFFLES_DECODE` in-shuffles.
-const CANONICAL_LENGTH: usize = 54;
+/// See <https://oeis.org/A002326/list> for a table of those values.
+const CANONICAL_LENGTH: usize = 64; // n = 32
 
-const SHUFFLES_ENCODE: usize = 18;
+const SHUFFLES_ENCODE: usize = 10;
 const SHUFFLES_DECODE: usize = 2;
 
 /// Zero-pads all human addresses to make them fit the canonical_length and
@@ -46,7 +48,7 @@ const SHUFFLES_DECODE: usize = 2;
 /// This is not really smart, but allows us to see a difference (and consistent length for canonical adddresses).
 #[derive(Copy, Clone)]
 pub struct MockApi {
-    /// Length of canonical addresses created with this API. Contracts should not make any assumtions
+    /// Length of canonical addresses created with this API. Contracts should not make any assumptions
     /// what this value is.
     canonical_length: usize,
     /// When set, all calls to the API fail with BackendError::Unknown containing this message
@@ -88,18 +90,20 @@ impl BackendApi for MockApi {
         }
 
         // Dummy input validation. This is more sophisticated for formats like bech32, where format and checksum are validated.
-        if normalized.len() < 3 {
+        let min_length = 3;
+        let max_length = self.canonical_length;
+        if normalized.len() < min_length {
             return (
                 Err(BackendError::user_err(
-                    "Invalid input: human address too short",
+                    format!("Invalid input: human address too short for this mock implementation (must be >= {min_length})."),
                 )),
                 gas_info,
             );
         }
-        if normalized.len() > self.canonical_length {
+        if normalized.len() > max_length {
             return (
                 Err(BackendError::user_err(
-                    "Invalid input: human address too long",
+                    format!("Invalid input: human address too long for this mock implementation (must be <= {max_length})."),
                 )),
                 gas_info,
             );
@@ -229,6 +233,13 @@ mod tests {
         let canonical = api.canonical_address(&original).0.unwrap();
         let recovered = api.human_address(&canonical).0.unwrap();
         assert_eq!(recovered, "cosmwasmchef");
+
+        // Long input (Juno contract address)
+        let original =
+            String::from("juno1v82su97skv6ucfqvuvswe0t5fph7pfsrtraxf0x33d8ylj5qnrysdvkc95");
+        let canonical = api.canonical_address(&original).0.unwrap();
+        let recovered = api.human_address(&canonical).0.unwrap();
+        assert_eq!(recovered, original);
     }
 
     #[test]
@@ -247,7 +258,7 @@ mod tests {
         let api = MockApi::default();
         let human = "1";
         match api.canonical_address(human).0.unwrap_err() {
-            BackendError::UserErr { .. } => {}
+            BackendError::UserErr { msg } => assert!(msg.contains("too short")),
             err => panic!("Unexpected error: {:?}", err),
         }
     }
@@ -257,7 +268,7 @@ mod tests {
         let api = MockApi::default();
         let human = "longer-than-the-address-length-supported-by-this-api-longer-than-54";
         match api.canonical_address(human).0.unwrap_err() {
-            BackendError::UserErr { .. } => {}
+            BackendError::UserErr { msg } => assert!(msg.contains("too long")),
             err => panic!("Unexpected error: {:?}", err),
         }
     }